First pass RenderSystem code migration
RenderSceneGraph code is a proxy change to get the code out of RenderSystem and into something else, and RenderSystem manages the different constituent systems to drive what it means to be a scene for rendering purposes in RZE. As a result, none of the code in RenderSceneGraph should be considered reliable. At all. It will all go, API and all. This change is just a stop gap to get it moving in the right direction. It's gross and buggy if you use it anywhere outside of just keeping things working.

Also added a brain dump txt wrt render pipeline responsibilities in RZE