fixing game rendering
i broke the game rendering when i fixed the editor aspect ratio issue. since the editor sets the viewport whenever it draws to SceneViewPanel, the fix worked for editor, but the game never did so. fix was to set the viewport size whenever we resize the window. the code is kinda wonky in that if you end up in a situation where you set the viewport, then resize the window, you'll need to re-set your viewport. i will need to address this in the future.