RenderObject stores pointer to buffers now
Always wanted to remove the vector-based indexing reference system and now ownership lies solely on the RenderObject to deal with it's resources. this will eventually unlock a much less complex deletion scheme for renderobjects and i can then re-implement functionality and move on to more rendering tasks. this is a checkpoint commit and doesnt have proper resource reclamation by RenderObject so resources will be leaking in situations (but it probably crashes anyway since thats what im trying to fix)